Updating log code to only call log if loging for the appropriate level is enabled
git-svn-id: https://svn.apache.org/repos/asf/incubator/tuscany/java/das@547544 13f79535-47bb-0310-9956-ffa450edef68

+++ samples/dbconfig/src/main/java/org/apache/tuscany/das/rdb/dbconfig/DBDataHelper.java
@@ -33,25 +33,33 @@ import org.apache.log4j.Logger;
 public class DBDataHelper {
     private static final String CLASS_NAME = "DBDataHelper";
 
-    private final Logger logger = Logger.getLogger("DBDataHelper.class");
+    private final Logger logger = Logger.getLogger(DBDataHelper.class);
 
     private final DBConfig dbConfig;
 
     protected DBDataHelper(DBConfig dbConfig) {
-        this.logger.log(Level.DEBUG, "DBDataHelper()");
+        if (logger.isDebugEnabled()) {
+            this.logger.log(Level.DEBUG, "DBDataHelper()");
+        }
+        
         this.dbConfig = dbConfig;
     }
 
     public boolean isDatabasePopulated() {
         boolean isPopulated = true;
 
-        this.logger.log(Level.DEBUG, CLASS_NAME + ".isDatabasePopulated()");
+        if (logger.isDebugEnabled()) {
+            this.logger.log(Level.DEBUG, CLASS_NAME + ".isDatabasePopulated()");
+        }
 
         Iterator tableIterator = dbConfig.getTable().iterator();
         while (tableIterator.hasNext()) {
             Table table = (Table) tableIterator.next();
 
-            this.logger.log(Level.DEBUG, CLASS_NAME + ".isDatabasePopulated() calling isTablePopulated() for '" + table.getName() + "'");
+            if (logger.isDebugEnabled()) {
+                this.logger.log(Level.DEBUG, CLASS_NAME + ".isDatabasePopulated() calling isTablePopulated() for '" + table.getName() + "'");
+            }
+            
             isPopulated = this.isTablePopulated(table.getName());
             if (isPopulated == false) {
                 break;
@@ -81,11 +89,16 @@ public class DBDataHelper {
             rs.next();
 
             if (rs != null) {
-                this.logger.log(Level.DEBUG, CLASS_NAME + ".isTablePopulated()=> pointer set");
+                if (logger.isDebugEnabled()) {
+                    this.logger.log(Level.DEBUG, CLASS_NAME + ".isTablePopulated()=> pointer set");
+                }
             }
 
             int count = rs.getInt(1);
-            this.logger.log(Level.DEBUG, CLASS_NAME + ".isTablePopulated()=> '" + tableName + "' => " + count);
+            
+            if (logger.isDebugEnabled()) {
+                this.logger.log(Level.DEBUG, CLASS_NAME + ".isTablePopulated()=> '" + tableName + "' => " + count);
+            }
 
             if (count > 0) {
                 isPopulated = true;
@@ -125,13 +138,17 @@ public class DBDataHelper {
               sqlBuffer.append(columnName).append(",");
           	}
           	else{
-          		this.logger.log(Level.DEBUG, CLASS_NAME + ".initializeDatabaseData()=> auto increment column => " + i + " "+columnName);
+                if (logger.isDebugEnabled()) {
+                    this.logger.log(Level.DEBUG, CLASS_NAME + ".initializeDatabaseData()=> auto increment column => " + i + " "+columnName);
+                }
           	}
         }
         sqlBuffer.append(rsMetaData.getColumnName(i)).append(") ");
         sqlBuffer.append("VALUES (").append(rowValues).append(")");
-        
-        this.logger.log(Level.DEBUG, CLASS_NAME + ".initializeDatabaseData()=> SQL Clause => " + sqlBuffer);
+
+        if (logger.isDebugEnabled()) {
+            this.logger.log(Level.DEBUG, CLASS_NAME + ".initializeDatabaseData()=> SQL Clause => " + sqlBuffer);
+        }
         
         return sqlBuffer.toString();
     }
@@ -147,7 +164,10 @@ public class DBDataHelper {
             Iterator tableIterator = dbConfig.getTable().iterator();
             while (tableIterator.hasNext()) {
                 Table table = (Table) tableIterator.next();
-                this.logger.log(Level.DEBUG, CLASS_NAME + ".initializeDatabaseData()=> INSERT FOR TABLE => '" + table.getName() + "'");
+                
+                if (logger.isDebugEnabled()) {
+                    this.logger.log(Level.DEBUG, CLASS_NAME + ".initializeDatabaseData()=> INSERT FOR TABLE => '" + table.getName() + "'");
+                }
          
                 //String columnClause = generateInsertSQL(dbConnection, table.getName());
                                 
@@ -157,20 +177,26 @@ public class DBDataHelper {
                     String tableRow = (String) dataIterator.next();                    
                     String sqlString = generateInsertSQL(dbConnection, tableName, tableRow);
 
-                    this.logger.log(Level.DEBUG, CLASS_NAME + ".initializeDatabaseData()=> sqlString => '" + sqlString + "'");
+                    if (logger.isDebugEnabled()) {
+                        this.logger.log(Level.DEBUG, CLASS_NAME + ".initializeDatabaseData()=> sqlString => '" + sqlString + "'");
+                    }
 
                     try {
                         dbStatement.executeUpdate(sqlString);
                     }catch(SQLException e){
                     	//e.printStackTrace();
                         //ignore and jump to new table
-                        this.logger.log(Level.DEBUG, CLASS_NAME + ".initializeDatabaseData() - Error inserting table data : " + e.getMessage(), e);
+                        if (logger.isDebugEnabled()) {
+                            this.logger.log(Level.DEBUG, CLASS_NAME + ".initializeDatabaseData() - Error inserting table data : " + e.getMessage(), e);
+                        }
                     }
                 }
             }
         } catch (SQLException e) {
             // ignore and return false
-            this.logger.log(Level.DEBUG, CLASS_NAME + ".initializeDatabaseData() - Internal error : " +  e.getMessage(), e);
+            if (logger.isDebugEnabled()) {
+                this.logger.log(Level.DEBUG, CLASS_NAME + ".initializeDatabaseData() - Internal error : " +  e.getMessage(), e);
+            }
         } finally {
             try {
                 dbStatement.close();
@@ -193,20 +219,26 @@ public class DBDataHelper {
             for(int i=tables.size()-1; i>-1; i-- ){
                 Table table = (Table) tables.get(i);
                 String sqlString = "DELETE FROM " + table.getName() ;
-                
-               this.logger.log(Level.DEBUG, CLASS_NAME + ".deleteDatabaseData()=> sqlString => '" + sqlString + "'");
+               
+                if (logger.isDebugEnabled()) {
+                    this.logger.log(Level.DEBUG, CLASS_NAME + ".deleteDatabaseData()=> sqlString => '" + sqlString + "'");
+                }
 
                 try {
                     dbStatement.executeQuery(sqlString);
                 }catch(SQLException e){
                 	e.printStackTrace();
                     //ignore and jump to new table
-                    this.logger.log(Level.DEBUG, CLASS_NAME + ".deleteDatabaseData() - Error inserting table data : " + e.getMessage(), e);
+                    if (logger.isDebugEnabled()) {
+                        this.logger.log(Level.DEBUG, CLASS_NAME + ".deleteDatabaseData() - Error inserting table data : " + e.getMessage(), e);
+                    }
                 }            	            
             }
         } catch (SQLException e) {
             // ignore and return false
-            this.logger.log(Level.DEBUG, CLASS_NAME + ".initializeDatabaseData() - Internal error : " +  e.getMessage(), e);
+            if (logger.isDebugEnabled()) {
+                this.logger.log(Level.DEBUG, CLASS_NAME + ".initializeDatabaseData() - Internal error : " +  e.getMessage(), e);
+            }
         } finally {
             try {
                 dbStatement.close();
